backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
nltk==3.8.1
python-docx==1.1.0
aiofiles==23.2.1
requests==2.31.0
torch==2.2.0
transformers==4.37.2
sentencepiece==0.1.99
